Added vehicle agent.
